feat: Rewrite genmap.php to genmap.py, add WKD hash entry
This commit rewrites the genmap.php to genmap.py and adds a "W" entry meaning a WKD hash entry.
This commit also adds requirements.txt for pip install -r requirements.txt

The WKD hash is an SHA-1 hash with a Z-Base-32 encoded digest of the local part of an email address.
This will be used to serve https://gnuweeb.org/.well-known/openpgpkey/<hash> by the server later.

The server will first read the map.txt, filter only the gnuweeb.org email address, and then serve the public keys.
For automatic map.txt updates, we can use something like GitHub Actions to run the genmap.py for every change in the keys directory and then push the new map.txt to the repo.
The server will also have a route for webhook where we can utilize GitHub webhook to trigger the server to add or update served public keys.

I will need your opinions on this.

REF: https://datatracker.ietf.org/doc/draft-koch-openpgp-webkey-service

Signed-off-by: Hazmi Alfarizqi <contact@hzmi.xyz>